Relocate now JVM-only tests
